Modularização das funções e componentização dos Todo's
